feature panel should not get closed when flip the visulisation, rename visulisation flip to back
