+++ src/main/webapp/app/modules/canvas/visualisation/widgets/visualisation-back.tsx
